Move Prisma-singleton to /lib
