+++ src/components/Buttons/ButtonStyles.module.css
@@ -1,45 +1,68 @@
-.primaryButton {
-    width: fit-content;
-    padding: 0.3rem 0.6rem; 
-    background: #c6e3ff;
-    color: var(--text-primary);
-    font-size: 1.3rem;
-    font-weight: 500;
-    border-radius: 4px;
-    border: 1px solid #5aacf9;  
-    cursor: pointer;
-    transition: background-color 0.3s ease, box-shadow 0.3s ease;
-  }
-  
-  .primaryButton:hover,  .primaryButton:focus {
-    background: #7dbefc; 
-    box-shadow: 0 0 0 3px rgba(214, 234, 255, 0.7); 
-  }
-  
-  .primaryButton:active {
-    background: #5aacf9; 
-    box-shadow: inset 0 0 5px rgba(0, 0, 0, 0.2);
-  }
+.button {
+  width: fit-content;
+  padding: 0.3rem 0.6rem;
+  color: var(--text-primary);
+  font-size: 1.1rem;
+  font-weight: 500;
+  cursor: pointer;
+  transition: all 0.3s ease-in-out;
+}
 
-  .secondaryButton {
-    width: fit-content;
-    padding: 0.3rem 0.6rem; 
-    background: #f0f0f0;
-    color: var(--text-primary);
-    font-size: 1.3rem;
-    font-weight: 500;
-    border-radius: 4px;
-    border: 1px solid #cccccc;  
-    cursor: pointer;
-    transition: background-color 0.3s ease, box-shadow 0.3s ease;
-  }
-  
-  .secondaryButton:hover, .secondaryButton:focus {
-    background: #e0e0e0; 
-    box-shadow: 0 0 0 3px rgba(200, 200, 200, 0.7); 
-  }
-  
-  .secondaryButton:active {
-    background: #cccccc; 
-    box-shadow: inset 0 0 5px rgba(0, 0, 0, 0.2);
-  }
+button.button:disabled {
+  background: #5e5e5e;
+  border: 1px solid #5e5e5e;
+}
+
+.button.primary {
+  background: #c6e3ff;
+  border-radius: 4px;
+  border: 1px solid #5aacf9;
+}
+
+.button.primary:hover, .button.primary:focus {
+  background: #7dbefc;
+  box-shadow: 0 0 0 3px rgba(200, 200, 200, 0.7);
+}
+
+.button.primary:active {
+  background: #5aacf9;
+  box-shadow: inset 0 0 5px rgba(0, 0, 0, 0.2);
+}
+
+.button.secondary {
+  background: #f0f0f0;
+  border-radius: 4px;
+  border: 1px solid #cccccc;
+}
+
+.button.secondary:hover, .button.secondary:focus {
+  background: #e0e0e0;
+  box-shadow: 0 0 0 3px rgba(200, 200, 200, 0.7);
+}
+
+.button.secondary:active {
+  background: #cccccc;
+  box-shadow: inset 0 0 5px rgba(0, 0, 0, 0.2);
+}
+
+.button.transparent {
+  border: none;
+  outline: none;
+}
+
+.button.warning {
+  border-color: hsl(0, 100%, 70%);
+  border-radius: 4px;
+  background: hsl(0, 100%, 70%);
+}
+
+.button.warning:hover, .button.secondary:focus {
+  background: hsl(0, 100%, 65%);
+  box-shadow: 0 0 0 3px hsl(0, 100%, 65%, 0.7);
+}
+
+.button.warning:active {
+  background: hsl(0, 100%, 65%);
+  box-shadow: inset 0 0 5px rgba(0, 0, 0, 0.2);
+}
+
