--- src/db/projects.ts
@@ -1,19 +1,0 @@
-import prisma from "./db"
-
-export type Project = {
-  title: string
-  video: string
-  gitHub: string
-  mainSection: {
-    image: { src: string; alt: string }
-    intro: string
-    learnings: string
-  }
-  techStack: string[]
-  features: string[]
-  
-}
-
-export async function getProjects() {
-  return await prisma.project.findMany()
-}
